# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api",
        "description": "RESTful API using Laravel",
        "contact": {
            "email": "dunierf@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get roles collection",
                "description": "Get a list of roles",
                "operationId": "77e3636d8592997ce9d6d27b50857a69",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoleDtoResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get a role",
                "description": "Get a role by id",
                "operationId": "7fb0667a295b7678a30353aa11739683",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDtoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RoleDtoResource": {
                "title": "Role",
                "description": "Role Dto",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "int64",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    }
                },
                "type": "object"
            }
        }
    }
}